Move check_tutorials_setup method
