+++ examples/tutorials/analysis-1d/cta_sensitivity.py
@@ -47,7 +47,7 @@ from gammapy.maps import MapAxis, RegionGeom
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.check import check_tutorials_setup
 
 check_tutorials_setup()
