+++ examples/tutorials/analysis-1d/extended_source_spectral_analysis.py
@@ -106,7 +106,7 @@ from gammapy.makers import (
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
