+++ examples/tutorials/analysis-1d/spectral_analysis.py
@@ -124,7 +124,7 @@ from gammapy.visualization import plot_spectrum_datasets_off_regions
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
