+++ examples/tutorials/analysis-1d/spectral_analysis_hli.py
@@ -80,7 +80,7 @@ from gammapy.modeling.models import Models, SkyModel
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
