+++ examples/tutorials/analysis-1d/spectral_analysis_rad_max.py
@@ -137,7 +137,7 @@ from gammapy.visualization import plot_spectrum_datasets_off_regions
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
