+++ examples/tutorials/analysis-1d/spectrum_simulation.py
@@ -74,7 +74,7 @@ from gammapy.maps import MapAxis, RegionGeom
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
