+++ examples/tutorials/analysis-2d/detect.py
@@ -69,7 +69,7 @@ import numpy as np
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
