+++ examples/tutorials/analysis-3d/analysis_3d.py
@@ -38,7 +38,7 @@ from gammapy.datasets import MapDataset
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
