+++ examples/tutorials/analysis-3d/analysis_mwl.py
@@ -65,7 +65,7 @@ from pathlib import Path
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
