+++ examples/tutorials/analysis-3d/cta_data_analysis.py
@@ -66,7 +66,7 @@ log.setLevel(logging.ERROR)
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
