+++ examples/tutorials/analysis-3d/event_sampling.py
@@ -111,7 +111,7 @@ from gammapy.modeling.models import (
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
