+++ examples/tutorials/analysis-3d/flux_profiles.py
@@ -59,7 +59,7 @@ from gammapy.modeling.models import PowerLawSpectralModel
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
