+++ examples/tutorials/analysis-time/light_curve.py
@@ -69,7 +69,7 @@ log = logging.getLogger(__name__)
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
