+++ examples/tutorials/analysis-time/light_curve_flare.py
@@ -88,7 +88,7 @@ from gammapy.makers import (
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
