+++ examples/tutorials/analysis-time/light_curve_simulation.py
@@ -96,7 +96,7 @@ from gammapy.data import observatory_locations
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
