+++ examples/tutorials/analysis-time/pulsar_analysis.py
@@ -54,7 +54,7 @@ from gammapy.estimators import FluxPointsEstimator
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
