+++ examples/tutorials/api/astro_dark_matter.py
@@ -47,7 +47,7 @@ import matplotlib.pyplot as plt
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
