+++ examples/tutorials/api/catalog.py
@@ -58,7 +58,7 @@ from gammapy.catalog import CATALOG_REGISTRY
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
