+++ examples/tutorials/api/datasets.py
@@ -48,7 +48,7 @@ from gammapy.utils.scripts import make_path
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
