+++ examples/tutorials/api/fitting.py
@@ -46,7 +46,7 @@ from itertools import combinations
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
