+++ examples/tutorials/api/makers.py
@@ -37,7 +37,7 @@ import numpy as np
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
