+++ examples/tutorials/api/maps.py
@@ -62,7 +62,7 @@ from gammapy.data import EventList
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
