+++ examples/tutorials/api/mask_maps.py
@@ -104,7 +104,7 @@ from gammapy.modeling.models import FoVBackgroundModel
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
