+++ examples/tutorials/api/model_management.py
@@ -77,7 +77,7 @@ from regions import CircleSkyRegion
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
