+++ examples/tutorials/api/models.py
@@ -36,7 +36,7 @@ from gammapy.maps import Map, WcsGeom, MapAxis
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
