+++ examples/tutorials/data/cta.py
@@ -69,7 +69,7 @@ from gammapy.irf import EffectiveAreaTable2D, load_cta_irfs
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
