+++ examples/tutorials/data/hess.py
@@ -53,7 +53,7 @@ from gammapy.visualization import plot_theta_squared_table
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
