+++ examples/tutorials/starting/analysis_1.py
@@ -79,7 +79,7 @@ from gammapy.analysis import Analysis, AnalysisConfig
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
