+++ examples/tutorials/starting/analysis_2.py
@@ -91,7 +91,7 @@ from gammapy.modeling.models import (
 )
 from gammapy.modeling import Fit
 from gammapy.estimators import FluxPointsEstimator
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 ######################################################################
 # Check setup
