+++ examples/tutorials/starting/overview.py
@@ -78,7 +78,7 @@ from astropy.coordinates import SkyCoord
 ######################################################################
 # Check setup
 # -----------
-from gammapy.utils.docs import check_tutorials_setup
+from gammapy.utils.checks import check_tutorials_setup
 
 check_tutorials_setup()
